refactor: move things and group things
